requirements/common.txt
pillow>=11.0.0,<12
tkfontawesome>=